Edit: Moved shared providers implementations to infra layer
