+++ src/shared/container/providers/HTTPProvider/index.ts
@@ -1,5 +1,5 @@
 import IHTTPProvider from 'shared/domain/providers/IHTTPProvider';
-import AxiosHTTPProvider from 'shared/container/providers/HTTPProvider/implementations/AxiosHTTPProvider';
+import AxiosHTTPProvider from 'shared/infra/providers/AxiosHTTPProvider';
 
 export default function makeHttpProvider(): IHTTPProvider {
   const httpProvider = new AxiosHTTPProvider();
