+++ src/shared/container/providers/JwtProvider/index.ts
@@ -1,5 +1,5 @@
 import { IJwtProvider } from 'shared/domain/providers/IJwtProvider';
-import JwtDecodeJwtProvider from './implementations/JwtDecodeJwtProvider';
+import JwtDecodeJwtProvider from 'shared/infra/providers/JwtDecodeJwtProvider';
 
 export default function makeJwtProvider(): IJwtProvider {
   const jwtProvider = new JwtDecodeJwtProvider();
