+++ src/shared/infra/providers/JwtDecodeJwtProvider.ts
@@ -1,4 +1,5 @@
 import decode from 'jwt-decode';
+
 import { IJwtProvider } from 'shared/domain/providers/IJwtProvider';
 
 export default class JsonwebtokenJwtProvider implements IJwtProvider {
